[Feature] upgrade seatunnel on flink connnector conf build.

+++ scaleph-core/scaleph-core-di/src/main/java/cn/sliew/scaleph/core/di/service/impl/DiClusterConfigServiceImpl.java
@@ -81,7 +81,7 @@ public class DiClusterConfigServiceImpl implements DiClusterConfigService {
     }
 
     @Override
-    public int totalCnt() {
+    public Long totalCnt() {
         return this.diClusterConfigMapper.selectCount(null);
     }
 }
